requirements.txt
fastapi
sqlalchemy
psycopg2
dotenv
uvicorn
passlib
pyjwt